+++ tests/plugins/dkim_sign.js
@@ -16,6 +16,7 @@ const _set_up = function (done) {
 
     this.connection = Connection.createConnection();
     this.connection.init_transaction();
+    this.connection.transaction.mail_from = {};
 
     done();
 }
@@ -26,7 +27,7 @@ exports.get_sender_domain = {
         test.expect(1);
         delete this.connection.transaction;
         test.equal(
-            this.plugin.get_sender_domain(this.connection.transaction),
+            this.plugin.get_sender_domain(this.connection),
             undefined
         );
         test.done();
@@ -34,7 +35,7 @@ exports.get_sender_domain = {
     'no headers': function (test) {
         test.expect(1);
         test.equal(
-            this.plugin.get_sender_domain(this.connection.transaction),
+            this.plugin.get_sender_domain(this.connection),
             undefined
         );
         test.done();
@@ -43,7 +44,7 @@ exports.get_sender_domain = {
         test.expect(1);
         this.connection.transaction.header.add('Date', utils.date_to_str(new Date()));
         test.equal(
-            this.plugin.get_sender_domain(this.connection.transaction),
+            this.plugin.get_sender_domain(this.connection),
             undefined
         );
         test.done();
@@ -51,21 +52,22 @@ exports.get_sender_domain = {
     'no from header, env MAIL FROM': function (test) {
         test.expect(1);
         this.connection.transaction.mail_from = new Address.Address('<test@example.com>');
-        const r = this.plugin.get_sender_domain(this.connection.transaction);
+        const r = this.plugin.get_sender_domain(this.connection);
         test.equal('example.com', r);
         test.done();
     },
     'env MAIL FROM, case insensitive': function (test) {
         test.expect(1);
         this.connection.transaction.mail_from = new Address.Address('<test@Example.cOm>');
-        const r = this.plugin.get_sender_domain(this.connection.transaction);
+        const r = this.plugin.get_sender_domain(this.connection);
         test.equal('example.com', r);
         test.done();
     },
     'From header not a fqdn': function (test) {
         test.expect(1);
         this.connection.transaction.header.add('From', 'root (Cron Daemon)');
-        this.plugin.get_key_dir(this.connection, function (err, dir) {
+        const r = this.plugin.get_sender_domain(this.connection);
+        this.plugin.get_key_dir(this.connection, r, function (err, dir) {
             test.equal(dir, undefined);
             test.done();
         });
@@ -73,28 +75,28 @@ exports.get_sender_domain = {
     'from header, simple': function (test) {
         test.expect(1);
         this.connection.transaction.header.add('From', 'John Doe <jdoe@example.com>');
-        const r = this.plugin.get_sender_domain(this.connection.transaction);
+        const r = this.plugin.get_sender_domain(this.connection);
         test.equal('example.com', r);
         test.done();
     },
     'from header, case insensitive': function (test) {
         test.expect(1);
         this.connection.transaction.header.add('From', 'John Doe <jdoe@Example.Com>');
-        const r = this.plugin.get_sender_domain(this.connection.transaction);
+        const r = this.plugin.get_sender_domain(this.connection);
         test.equal('example.com', r);
         test.done();
     },
     'from header, less simple': function (test) {
         test.expect(1);
         this.connection.transaction.header.add('From', '"Joe Q. Public" <john.q.public@example.com>');
-        const r = this.plugin.get_sender_domain(this.connection.transaction);
+        const r = this.plugin.get_sender_domain(this.connection);
         test.equal('example.com', r);
         test.done();
     },
     'from header, RFC 5322 odd': function (test) {
         test.expect(1);
         this.connection.transaction.header.add('From', 'Pete(A nice \\) chap) <pete(his account)@silly.test(his host)>');
-        const r = this.plugin.get_sender_domain(this.connection.transaction);
+        const r = this.plugin.get_sender_domain(this.connection);
         test.equal('silly.test', r);
         test.done();
     },
@@ -102,7 +104,7 @@ exports.get_sender_domain = {
         test.expect(1);
         this.connection.transaction.header.add('From', 'ben@example.com,carol@example.com');
         this.connection.transaction.header.add('Sender', 'dave@example.net');
-        const r = this.plugin.get_sender_domain(this.connection.transaction);
+        const r = this.plugin.get_sender_domain(this.connection);
         test.equal('example.net', r);
         test.done();
     },
@@ -112,7 +114,7 @@ exports.get_sender_domain = {
         // addr parser doesn't support the RFC 6854 Group Syntax
         this.connection.transaction.header.add('From', 'Managing Partners:ben@example.com,carol@example.com;');
         this.connection.transaction.header.add('Sender', 'dave@example.net');
-        const r = this.plugin.get_sender_domain(this.connection.transaction);
+        const r = this.plugin.get_sender_domain(this.connection);
         test.equal('example.net', r);
         test.done();
     },
@@ -136,8 +138,8 @@ exports.get_key_dir = {
     },
     'no transaction': function (test) {
         test.expect(2);
-        this.plugin.get_key_dir(this.connection, function (err, dir) {
-            test.equal(err, undefined);
+        this.plugin.get_key_dir(this.connection, '', function (err, dir) {
+            test.equal(err.message, 'missing domain');
             test.equal(dir, undefined);
             test.done();
         });
@@ -145,7 +147,7 @@ exports.get_key_dir = {
     'no key dir': function (test) {
         test.expect(1);
         this.connection.transaction.mail_from = new Address.Address('<matt@non-exist.com>');
-        this.plugin.get_key_dir(this.connection, function (err, dir) {
+        this.plugin.get_key_dir(this.connection, 'non-exist.com', function (err, dir) {
             test.equal(dir, undefined);
             test.done();
         });
@@ -154,7 +156,7 @@ exports.get_key_dir = {
         test.expect(1);
         process.env.HARAKA = path.resolve('tests');
         this.connection.transaction.mail_from = new Address.Address('<matt@example.com>');
-        this.plugin.get_key_dir(this.connection, function (err, dir) {
+        this.plugin.get_key_dir(this.connection, 'example.com', function (err, dir) {
             // console.log(arguments);
             const expected = path.resolve('tests','config','dkim','example.com');
             test.equal(dir, expected);