+++ Changes.md
@@ -1,9 +1,13 @@
 
 ## 2.8.22 - Mmm DD, 2018
 
-* New Features
-* Fixes
-* Changes
+### New Features
+### Fixes
+### Changes
+
+* dkim_sign: improved log messages #2499
+
+* ehlo_hello_message: config/ehlo_hello_message can be used to overwrite the EHLO/HELO msg replacing `, Haraka is at your service` #2498
 
 
 ## 2.8.21 - Jul 20, 2018