+++ Changes.md
@@ -1,10 +1,17 @@
 
 ## 2.8.22 - Mmm DD, 2018
 
-* New Features
-* Fixes
-* Changes
+### New Features
+### Fixes
+* connection: nove max_mime_part config load to connection init #2528
 
+### Changes
+
+* dkim_sign: improved log messages #2499
+
+* ehlo_hello_message: config/ehlo_hello_message can be used to overwrite the EHLO/HELO msg replacing `, Haraka is at your service` #2498
+
+* connection: add connection.remote.is_local flag for detecting loopback and link local IPs
 
 ## 2.8.21 - Jul 20, 2018