+++ Changes.md
@@ -1,5 +1,5 @@
 
-## 2.8.22 - Mmm DD, 2018
+## 2.8.22 - Nov DD, 2018
 
 ### New Features
 ### Fixes
@@ -16,6 +16,7 @@
 * ehlo_hello_message: config/ehlo_hello_message can be used to overwrite the EHLO/HELO msg replacing `, Haraka is at your service` #2498
 
 * connection: add connection.remote.is_local flag for detecting loopback and link local IPs
+
 
 ## 2.8.21 - Jul 20, 2018